Code cleanup + hasRolled saving
Separated parts of EMMItem into ModifierItem.
EMMPlayer renamed to ModifierPlayer.
hasRolled now saved and loaded.
